requirements.txt
aniso8601==9.0.1
black==23.3.0
blinker==1.6.2
-e git+https://github.com/SinclairPythonAkoto/blog_api@b8ed43728eb6d87159494cb1e7022e9c15be50a2#egg=blog_api
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
flask-sqlalchemy==3.0.5
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.17
tomli==2.0.1
typing-extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
zope.interface==6.0
